Version Prueba If {hay apocalipsis] return "Saquen a Chapa", else {"sigan viendo"}

+++ Root Out!/Assets/Scripts/Weapons/IBullet.cs
@@ -1,4 +1,4 @@
 public interface IBullet
 {
-    void SetDamage(int damageAmount);
+    void SetDamage(float damageAmount);
 }
